sciforum/requirements.txt
aiodns==2.0.0
aiofile==3.3.3
aiohttp==3.7.3
alabaster==0.7.12
argh==0.26.2
asgiref==3.2.10
astropy==4.0.2
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
Automat==20.2.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
bkcharts==0.2
blis==0.7.4
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
CacheControl==0.12.6
cachetools==4.2.0
caio==0.6.3
catalogue==1.0.0
cchardet==2.1.7
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
constantly==15.1.0
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
cytoolz==0.11.0
decorator==4.4.2
defusedxml==0.7.0rc1
dj-rest-auth==2.1.1
Django==3.0.5
django-allauth==0.42.0
django-braces==1.14.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-enumfields==2.0.0
django-extensions==3.0.3
django-filter==2.4.0
django-model-utils==4.0.0
django-mongodb-engine @ git+https://github.com/django-nonrel/mongodb-engine@1314786975d3d17fb1a2e870dfba004158a5e431
django-mysql==3.9.0
django-notifications-hq==1.6.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.0.7
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-word-filter==0.2.0
djangotoolbox @ git+https://github.com/django-nonrel/djangotoolbox@7e2c47e2ab5ee98d0c0f47c5ccce08e75feda48a
djongo==1.3.3
docutils==0.16
drf-writable-nested==0.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
firebase-admin==4.5.0
Flask==1.1.2
future==0.18.2
glob2==0.7
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
h5py==2.10.0
HeapDict==1.0.1
httplib2==0.18.1
hyperlink==21.0.0
idna==2.5
imagesize==1.2.0
incremental==17.5.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.5
jsonfield==3.1.0
jupyter==1.0.0
jupyter-core==4.6.3
jupyterlab==2.2.6
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.34.0
locket==0.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
mongoengine==0.22.1
mpmath==1.1.0
msgpack==1.0.0
multidict==5.1.0
multipledispatch==0.6.0
murmurhash==1.0.5
mysqlclient==2.0.1
nose==1.3.7
numba==0.51.2
numexpr==2.7.1
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.4
parsel==1.6.0
parso==0.7.0
partd==1.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.6.1
plac==1.1.3
pluggy==0.13.1
ply==3.11
preshed==3.0.5
prometheus-client==0.8.0
Protego==0.1.16
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycodestyle==2.6.0
pycosat==0.6.3
pycryptodomex==3.9.9
pycurl==7.43.0.6
PyDispatcher==2.0.5
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtPy==1.9.0
queuelib==1.5.0
requests-oauthlib==1.3.0
responses==0.10.8
rest-framework-generic-relations==2.0.0
rsa==4.6
ruamel-yaml==0.15.87
scikit-image==0.17.2
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
scrapy-splash==0.7.2
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.17.2
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.2.4
srsly==1.0.5
statsmodels==0.12.0
stream-chat==3.0.1
swapper==1.1.2.post1
tables==3.6.1
terminado==0.9.1
testpath==0.4.4
thinc==7.4.5
tifffile==2020.10.1
tornado==6.0.4
Twisted==20.3.0
unicodecsv==0.14.1
uritemplate==3.0.1
w3lib==1.22.0
wasabi==0.8.0
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.3
zict==2.0.0
zipp==3.1.0
zope.event==4.5.0